# src/main/java/t1/openschool/jwtdemo/dto/news/NewsResponseShortDto.java
package t1.openschool.jwtdemo.dto.news;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Response short news")
public class NewsResponseShortDto {
    private String header;
}