# src/main/java/t1/openschool/jwtdemo/dto/security/AuthRefreshRequestDto.java
package t1.openschool.jwtdemo.dto.security;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Token request by refreshToken")
public class AuthRefreshRequestDto {
    private String refreshToken;
}