# src/main/java/t1/openschool/jwtdemo/util/exceptions/handler/ApiError.java
package t1.openschool.jwtdemo.util.exceptions.handler;

import lombok.Builder;
import lombok.Data;
import org.springframework.http.HttpStatus;

@Data
@Builder
public class ApiError {
    private HttpStatus status;
    private String reason;
    private String message;
    private StackTraceElement[] errors;
}